Add files via upload
By using the feature of the doctors like experience we have  to predict their consultancy price that how we should pay. In this there are Many feature first the degree of a doctor that which level doctor are they. And other is the experience of that doctor and then the city in which they are doing is the most correlative eith their consultancy fees. By using that much feature we made models and check accuracy of that.